sortin listview, cleaned resultfragment

+++ app/src/main/java/com/example/elisarajaniemi/kellotusapp/ResultItem.java
@@ -11,7 +11,7 @@ public class ResultItem {
 
     public String address;
     public double time;
-    public double kellotusTime;
+    public int kellotusTime;
     public String comment;
     public String name;
     public int date;
@@ -19,7 +19,7 @@ public class ResultItem {
     public double longitude;
 
 
-    public ResultItem(String address, double time, double kellotusTime, String comment, String name, int date, double latitude, double longitude){
+    public ResultItem(String address, double time, int kellotusTime, String comment, String name, int date, double latitude, double longitude){
         this.address = address;
         this.time = time;
         this.kellotusTime = kellotusTime;
